+++ src/settings/ShowCapabilities/ShowCapabilities.js
@@ -1,8 +1,9 @@
 import React, { useState } from 'react';
 import { FormattedMessage } from 'react-intl';
 
-import { stripesConnect, useOkapiKy, useStripes } from '@folio/stripes/core';
+import { stripesConnect, useChunkedCQLFetch, useOkapiKy, useStripes } from '@folio/stripes/core';
 import { Button, Pane, Row, Col, SearchField, Select } from '@folio/stripes/components';
+import Capabilities from './Capabilities';
 
 const ShowCapabilities = () => {
   const SEARCH_BY_TYPES = {
@@ -43,10 +44,6 @@ const ShowCapabilities = () => {
     return searchIds;
   };
 
-  const lookUpPermissionDisplayNameById = (permissionName) => {
-    return stripes.discovery?.permissionDisplayNames?.[permissionName];
-  };
-
   const submit = async () => {
     if (searchBy === SEARCH_BY_TYPES.PERMISSION_DISPLAY_NAME) {
       const searchIds = searchForPermissionDisplayName(query);
@@ -55,33 +52,17 @@ const ShowCapabilities = () => {
       searchParams.query = `permission=*${query}*`;
     }
 
-    const capabilitiesResponse = await ky.get('capabilities', { searchParams }).json();
-    setCapabilitiesResults(capabilitiesResponse);
+    // const capabilitiesResponse = await ky.get('capabilities', { searchParams }).json();
+    // setCapabilitiesResults(capabilitiesResponse);
 
-    const capabilitySetsResponse = await ky.get('capability-sets', { searchParams }).json();
-    setCapabilitySetsResults(capabilitySetsResponse);
+    // const capabilitySetsResponse = await ky.get('capability-sets', { searchParams }).json();
+    // setCapabilitySetsResults(capabilitySetsResponse);
   };
 
   const handleKeyDown = (event) => {
     if (event.key === 'Enter') {
       submit();
     }
-  };
-
-  const displayList = (resultList, listType) => {
-    return resultList[listType]?.map((value) => (
-      <ul>
-        <li>{value.name}</li>
-        <ul>
-          <li><strong>type:</strong> {value.type}</li>
-          <li><strong>applicationId:</strong> {value.applicationId}</li>
-          <li><strong>resource:</strong> {value.resource}</li>
-          <li><strong>action:</strong> {value.action}</li>
-          <li><strong>permissionName:</strong> {value.permission}</li>
-          <li><strong>permissionDisplayName:</strong> {lookUpPermissionDisplayNameById(value.permission)}</li>
-        </ul>
-      </ul>
-    ));
   };
 
   return (
@@ -107,10 +88,11 @@ const ShowCapabilities = () => {
       </Row>
       <Row>
         <Col xs={12}>
-          { capabilitiesResults.capabilities?.length > 0 && <h3><FormattedMessage id="ui-developer.capabilities" /></h3> }
+          <Capabilities query={searchParams.query} />
+          {/* { capabilitiesResults.capabilities?.length > 0 && <h3><FormattedMessage id="ui-developer.capabilities" /></h3> }
           { displayList(capabilitiesResults, 'capabilities') }
           { capabilitySetsResults.capabilitySets?.length > 0 && <h3><FormattedMessage id="ui-developer.capabilitySets" /></h3> }
-          { displayList(capabilitySetsResults, 'capabilitySets') }
+          { displayList(capabilitySetsResults, 'capabilitySets') } */}
         </Col>
       </Row>
     </Pane>
